build(Gradle): Add a library conventions plugin to migrate build logic to
The use of `allprojects` and `subprojects` is discouraged [1] due to
configuration-time coupling between projects [2]. Prepare to avoid
these constructs by introducing a basic conventions plugin for ORT
Kotlin libraries to which shared build logic will be moved to in
upcoming changes.

The only libraries not considering by this change are all "clients"
libraries as well as "common-utils" and "spdx-utils" as these should
stay independent for eventually being moved out of the ORT code base.

[1]: https://docs.gradle.org/current/userguide/sharing_build_logic_between_subprojects.html#sec:convention_plugins_vs_cross_configuration
[2]: https://docs.gradle.org/current/userguide/multi_project_configuration_and_execution.html#sec:decoupled_projects

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>